Finished the TransferProcessorTest test cases. Going to create one final branch for output requirements.

+++ src/main/java/banking/Bank.java
@@ -36,7 +36,7 @@ public class Bank {
         new PassTime(accounts).passTheTime(months);
     }
 
-    public void transfer(Account accountFrom, Account accountTo, int amountToTransfer) {
+    public void transfer(Account accountFrom, Account accountTo, double amountToTransfer) {
         new Transfer(accountFrom, accountTo).transfer(amountToTransfer);
     }
 }
